requirements.txt
# Core dependencies for the assistive glasses system
openai>=1.0.0
requests>=2.25.0
Pillow>=9.0.0
gTTS>=2.3.0
pygame>=2.0.0

# AI and Speech Recognition dependencies
openai-whisper>=20231117
pyaudio>=0.2.11
numpy>=1.21.0

# Raspberry Pi specific dependencies
# Note: These should be installed separately on Raspberry Pi
# picamera2>=0.3.0
# RPi.GPIO>=0.7.0

# Optional dependencies for enhanced functionality
opencv-python>=4.5.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0 